+++ src/core/platforms/ctrader/CTraderAccount.ts
@@ -625,12 +625,14 @@ export class CTraderAccount extends MidaTradingAccount {
             uuid,
             connection: this.#connection,
             cTraderEmitter: this.#cTraderEmitter,
+            requestedProtection: directives.protection,
         });
 
         const plainSymbol: GenericObject = this.#symbols.get(symbol) as GenericObject;
         const completePlainSymbol: GenericObject = await this.#getCompletePlainSymbol(symbol);
         const lotUnits: number = Number(completePlainSymbol.lotSize) / 100;
         const normalizedVolume: number = requestedVolume * lotUnits * 100;
+        let setProtectionAfterExecution: boolean = false;
 
         requestDirectives = {
             symbolId: plainSymbol.symbolId.toString(),
@@ -672,10 +674,17 @@ export class CTraderAccount extends MidaTradingAccount {
                     requestDirectives.trailingStopLoss = true;
                 }
             }
+            else {
+                setProtectionAfterExecution = true;
+            }
         }
         else {
             requestDirectives.positionId = positionId;
             requestDirectives.orderType = "MARKET";
+
+            if (directives.protection) {
+                console.log("Order protection ignored, change the protection directly on the position");
+            }
         }
 
         const resolverEvents: string[] = directives.resolverEvents ?? [
@@ -718,6 +727,20 @@ export class CTraderAccount extends MidaTradingAccount {
                 this.#normalizedOrders.set(id, order);
             }
         });
+
+        // <set-protection-after-execution>
+        if (setProtectionAfterExecution) {
+            order.on("execute", async (): Promise<void> => {
+                const protection = directives.protection;
+                const openPosition: MidaPosition | undefined = (await this.getOpenPositions())
+                    .find((position: MidaPosition) => position.id === order.positionId);
+
+                if (openPosition && protection) {
+                    await openPosition.changeProtection(protection);
+                }
+            });
+        }
+        // </set-protection-after-execution>
 
         this.#sendCommand("ProtoOANewOrderReq", requestDirectives, uuid);