+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Features
+* Implement `isSymbolMarketOpen()` interface [#16](https://github.com/Reiryoku-Technologies/Mida-cTrader/pull/16)
+
 3.0.1 - 02-06-2022
 ===================
 ## Bug fixes